Add an in-memory cache in EventRepositoryImpl to speed things up
Earlier, due to multiple SQL queries which had to be run every time the
user starred or unstarred an event, the app feedback was pretty slow.

This cache is pretty simple and results in quick feedback when a user
stars or unstars an event.

Currently, it is only used when starring or unstarring events. Other
things(like filters) don't use it.