Extract the screen title widget and make the more screen
The screen title widget is used in all the top-level screens. So, it
makes sense to extract it out into a shared widget.

More screen's options are not clickable because their screens haven't
been made yet.

The screen is also not accessible right now because bottom nav hasn't
been updated to support navigation.